Move main page element to root of page folder
