--- src/pages/leadership/LeadershipLayout.jsx
@@ -1,10 +1,0 @@
-import { Outlet } from "react-router-dom";
-import { LeadershipProvider } from "./Components/LeadershipContext";
-
-export default function LeadershipLayout () {
-    return (
-        <LeadershipProvider>
-            <Outlet/>
-        </LeadershipProvider>
-    );
-}
